refactor: User Account -> Team
When a user signs up, an Account Request record is created, which is
used to create a Team when the user verifies their email.
An Account Request record is valid for 30 minutes after it is created.
Team can have Team Members, users can add other members from
the team page. The invited member can have their own account on
Frappe Cloud or not. Both type of users can be added to team.
If a user is a part of someone else's team, he can switch the team
from the team page.