fix: Handle errors during scheduling backups for multiple sites
Scenario: backup() is called for site_1 and site_2.
For site_1 a Site Backup is created and a Agent Job is created
create_http_request is deferred till commit.

For site_2 because there are already some pending backups "Too Many Backups" is thrown and the transaction is rolled back.

create_http_request now cannot be properly executed as
the Agent Job it is supposed to be called upon no longer exists.