fix: apply tax on amount due and not total
Take an example of an invoice where the total amount is 100. User got 20 prepaid credits.
Current logic will apply 18% tax on total so it becomes 118. Then credits are applied, amount due becomes 98.
But User paid 18% tax on the 20 amount. So, correct logic is to first apply credits and then apply tax on remaining amount. So amount due is 80, and with tax it becomes 94.4. So the user has to pay 94.4 and not 98.