fix: show Builds Suspended banner
Long sidenote:

I'm a bit pissed cause it's nearly 2130 on a Friday and I had to deal
with dashboard/src2. I wouldn't have been pissed had the API design of
dashboard/src2 not been dogshit.

If you went through the commit and are wondering why the AlertBanner is
being set in DetailPage.vue i.e. a "building block" component look no
further than src2/objects/bench.js and ask yourself where and how in
this ~1100 page file does one display a banner on the Benches page for
when builds have been suspended.

Chances are unless you're one of the authors, you won't have an answer.
And that is because this API is a poor attempt at declarative
programming. One that has been built on another poor attempt at
declarative programming.

By which I of course mean frappe-ui, not the components, but
specifically the part called "resources". An unecessarily convoluted and
tasteless way of interacting with FF. I mean ffs—onSuccess, onError—what
year is this? Even the language moved on from callbacks back in like
2017. And even if you want to deal with callbacks, literally
fetch(...).then(...).catch(...) is simpler than whatever is going
on with resources. Ugh.

End rant, I'm now going to have my long awaited dinner.