fix: send title instead of message in mail
When a deploy fails the press notification message set is
meant to be viewed in the UI and not in the users inbox.

This is my fault for overloading Press Notifications to
handle User Addressable failures, should have listened
to Faris on this.

That being said the title in most cases suffices as a message
and the email carries a link to the Deploy where the user
can view what has gone wrong in detail