feat(volumes): Set MariaDB to depend on mounts when data volumes are used
If MariaDB gets activated, then mounts will be activated as well.
If one of the mounts fails to activate, MariaDB will not be started.
If one of the mounts is explicitly stopped (or restarted), MariaDB will be stopped (or restarted).

If one of the mounts are stopped, MariaDB will be stopped too.

The mounts strictly have to be in active state for MariaDB to also be in active state.
Reference: https://www.freedesktop.org/software/systemd/man/latest/systemd.unit.html#BindsTo=