fix(saas-pool): Consider pending sites in standby count
While checking whether new standby site is required or not
We only check for active sites.
Now, it can happen that some server has lot of pending jobs
and New Site jobs waiting to be executed.

In that case, if we check for active sites, it will keep adding new sites